requirements.txt
torch>=2.0
pydra-config
matplotlib